// src/types/NavRefs.ts
import {MutableRefObject} from "react";

export type NavRefs = {
    start: MutableRefObject<any>;
    techStack: MutableRefObject<any>;
    projects: MutableRefObject<any>;
    activity: MutableRefObject<any>;
    career: MutableRefObject<any>;
    about: MutableRefObject<any>;
    contact: MutableRefObject<any>;
}